[Feat] 로또 수익률 계산

+++ src/main/java/lotto/domain/LottoRank.java
@@ -31,6 +31,6 @@ public enum LottoRank {
 	}
 
 	public long getTotal(int count) {
-		return count * prize;
+		return (long) count * (long) prize;
 	}
 }
